+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The template version should stay in sync with the engine version
 - Script `deployer` renamed to `deploy`
 - i18n auto-translator
 - Engine version 4.0.0
+- Bugfix with dialogue bleed
 
 ## 3.0.0 - 7/15/25